fix : 게시물 수정,삭제시 본인이 생성한 게시물만 삭제 가능하도록 변경
PostService에서 유저의 상태와 현제 접속한 유저의 id와 post의 userId를 비교하여 틀리면 예외문구를 출력하도록 변경

+++ src/main/java/com/example/newsfeed/service/PostService.java
@@ -12,11 +12,13 @@ import jakarta.servlet.http.HttpSession;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.PageRequest;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 import org.springframework.data.domain.Pageable;
 import org.springframework.web.server.ResponseStatusException;
 
 import java.util.List;
+import java.util.Objects;
 
 
 @Service
@@ -26,9 +28,11 @@ public class PostService {
     private final PostRepository postRepository;
     private final UserRepository userRepository;
 
+
     @Transactional
     public PostResponseDto createPost(PostRequestDto postRequestDto, HttpServletRequest request) {
         User user = findusers(request);
+
         Post post = new Post(postRequestDto.getContent(), postRequestDto.getContent(),user);
 
         Post savePost = postRepository.save(post);
@@ -45,7 +49,11 @@ public class PostService {
     }
 
     @Transactional
-    public PostResponseDto findById(Long postId) {
+    public PostResponseDto findById(Long postId, HttpServletRequest request) {
+        User user = findusers(request);
+        if (user.getUserStatus().equals("N")) {
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "존재하지 않는 게시물입니다.");
+        }
         return PostResponseDto.toDto(findPostById(postId));
     }
 
@@ -56,10 +64,12 @@ public class PostService {
 
     @Transactional
     public PostResponseDto updatePost(Long postId, PostUpdateRequestDto updateRequestDto, HttpServletRequest request) {
-        HttpSession session = request.getSession();
-
-
+        User user = findusers(request);
         Post post = findPostById(postId);
+        if (user.getUserStatus().equals("N") || !(user.getUserId().equals(post.getUser().getUserId()))) {
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "본인의 게시물만 수정할 수 있습니다.");
+        }
+
         if (updateRequestDto.getTitle() != null && updateRequestDto.getContent() != null) {
             post.update(updateRequestDto.getTitle(), updateRequestDto.getContent());
         } else if (updateRequestDto.getTitle() != null) {
@@ -72,12 +82,16 @@ public class PostService {
 
     @Transactional
     public void deletePost(Long postId, HttpServletRequest request) {
+        User user = findusers(request);
         Post post = findPostById(postId);
+        if (user.getUserStatus().equals("N") || !(user.getUserId().equals(post.getUser().getUserId()))) {
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "본인의 게시물만 삭제할 수 있습니다.");
+        }
         postRepository.deleteById(postId);
     }
 
     public User findusers(HttpServletRequest request) {
-        HttpSession session = request.getSession();
+        HttpSession session = request.getSession(false);
         Long userId = (Long) session.getAttribute("SESSION_KEY");
         User user = userRepository.findByIdOrElseThrow(userId);
         return user;
